+++ main.cpp
@@ -10,20 +10,20 @@
 
 using namespace std;
 
-glm::vec3 g_array[ARRAY_SIZE*ARRAY_SIZE*2] = { glm::vec3(0.0f) };
+glm::dvec3 g_array[ARRAY_SIZE*ARRAY_SIZE*2] = { glm::dvec3(0.0f) };
 mutex g_posMutex;
 
 int main()
 {
-    vector<glm::vec3 *> g_inst_positions;
+    vector<glm::dvec3 *> g_inst_positions;
 
     //create array and construct vector from it
     for(int i=0; i<ARRAY_SIZE; ++i) {
         for(int j=0; j<ARRAY_SIZE; ++j) {
             float x = (float)(-ARRAY_SIZE/2+j);
             float y = (float)(-ARRAY_SIZE/2+i);
-            g_array[2*(i*ARRAY_SIZE+j)] = glm::vec3(x, y, 0.0f);
-            g_array[2*(i*ARRAY_SIZE+j)+1] = glm::vec3(1.0f, 1.0f, 0.0f);
+            g_array[2*(i*ARRAY_SIZE+j)] = glm::dvec3(x, y, 0.0f);
+            g_array[2*(i*ARRAY_SIZE+j)+1] = glm::dvec3(1.0f, 1.0f, 0.0f);
         }
     }
     for(int i=0; i<sizeof(g_array)/sizeof(g_array[0]); i++)
@@ -43,7 +43,7 @@ int main()
         //update positions
         g_posMutex.lock();
         for (auto it = g_inst_positions.begin(); it != g_inst_positions.end(); it+=2) {
-            **it += glm::vec3(0.8f, 0.0f, 0.0f);
+            **it += glm::dvec3(0.8f, 0.0f, 0.0f);
         }
         g_posMutex.unlock();
